+++ NEWS.md
@@ -1,3 +1,12 @@
+## Imagewmark NEXT:
+
+### Changed:
+* Replaced OpenCV with libvips for the C++ watermark embedding pipeline
+* Moved to affine transform with bicubic interpolation for image resizing
+* Optimized PNG compression options when saving embedded images
+* Updated build dependencies to require libvips-dev instead of libopencv-dev
+
+
 ## Imagewmark 0.6.0:
 
 ### Added:
@@ -36,7 +45,7 @@
 * Fixed tests/Makefile: properly use cornersync mode from Makefile variable
 	https://github.com/tim-janik/imagewmark/issues/9
 * Fixed tests/gen-tests-mk: fix random size aspect ratio generation
-	https://github.com/tim-janik/imagewmark/issues/8, #8
+	https://github.com/tim-janik/imagewmark/issues/8
 * Fixed Enforce random aspect ratio in test suite
 	https://github.com/tim-janik/imagewmark/issues/6
 * Fixed gen-tests-mk: Makefile triggers /bin/bash: Argument list too long